// front/src/components/MpDraggable/config/mpDraggable.enum.ts
export enum MovingButtonButtonsVerticalArgs {
  up = 'up',
  down = 'down',
  last = 'last',
  first = 'first',
}

export enum MovingButtonContentVertical {
  first = '↑↑',
  up = '↑',
  down = '↓',
  last = '↓↓',
}

export enum MovingButtonButtonsHorizontalArgs {
  right = 'right',
  rightAll = 'rightAll',
  left = 'left',
  leftAll = 'leftAll',
}

export enum MovingButtonContentHorizontal {
  right = '›',
  left = '‹',
  rightAll = '«',
  leftAll = '»',
}